LocalizedEnumValue as interface #1053
